Deduplicator will now expose a channel containing metadata about photos and whether they are duplicates are not. Also some naming conventions have been changed.
The Serve() funciton will start the deduplication data gathering process. Serve is currently a blocking call and will not return until all photos have been proccessed and placed onto the output channel. This requires the channel to be of sufficient size to contain all the details about all photos processed. This should be modified to make it non blocking through the use of go routines